Polish: rename or limit scope of utility classes
