+++ entities/src/test/java/net/kemitix/naolo/entities/PetTest.java
@@ -31,7 +31,7 @@ class PetTest implements WithAssertions {
 
     @Provide
     static Arbitrary<ZonedDateTime> zonedDateTime() {
-        return ArbitraryProviders.zonedDateTime();
+        return ArbitrariesHelper.zonedDateTime();
     }
 
 }
